refactor(watch): extract pure channel logic into channelEngine, drop dead hasRecentInventory
