+++ src/renderer/shared/hooks/watch/index.ts
@@ -1,4 +1,5 @@
 export * from "./channelAllowlist";
+export * from "./channelEngine";
 export * from "./useChannels";
 export * from "./useWatchingActions";
 export * from "./useWatchingController";
